+++ lib/avromatic/model/builder.rb
@@ -5,6 +5,7 @@ require 'avromatic/model/configuration'
 require 'avromatic/model/value_object'
 require 'avromatic/model/configurable'
 require 'avromatic/model/attributes'
+require 'avromatic/model/serialization'
 require 'avromatic/model/messaging'
 
 module Avromatic
@@ -41,6 +42,7 @@ module Avromatic
           Avromatic::Model::Configurable,
           Avromatic::Model::Attributes,
           Avromatic::Model::ValueObject,
+          Avromatic::Model::Serialization,
           Avromatic::Model::Messaging
         ]
       end